Move capability tests out of regression_test.go into run_test.go
regression_test.go's contract is to reproduce bugs in the deleted
pre-rewrite engine. Three tests didn't fit: the ctx-aware-producer leak
test is self-labeled a capability test, and the two contract-violating
"engine still returns" tests characterize invariants of code the rewrite
introduced (the ctx exit path, the transformer channel stage) — neither
regresses anything the old engine had. Move them next to the other
RunPipeline behavior tests. Also note in the header why
Test_ErrorAfterDataClose_IsDelivered stays despite guarding a new-engine
fix rather than a 5c9985f one.
