+++ core/regression_test.go
@@ -10,8 +10,6 @@ import (
 	"time"
 
 	"github.com/psyduck-etl/sdk"
-
-	"github.com/gastrodon/psyduck/stdlib/flow"
 )
 
 // This file holds regression tests for the concurrency bugs the pre-rewrite
@@ -45,6 +43,17 @@ import (
 // #8 (produce-from seed closing without sending) is a parse-layer bug, not
 // a core-engine one; its regression test lives in
 // parse/hcl/hcl_test.go:TestParseProduceFromClosedSeed.
+//
+// One test here — Test_ErrorAfterDataClose_IsDelivered — guards a dropped-error
+// bug in the *new* engine's error forwarder that was found and fixed during the
+// rewrite, not in the 5c9985f engine reproduced above. It has no issue number
+// and isn't one of #10/#11/#12/#19, but it's a genuine regression against a
+// real fix, so it lives with the rest.
+//
+// Capability/invariant tests that exercise behavior the rewrite *introduced*
+// (the ctx-aware exit path, the transformer channel stage) rather than
+// reproducing a deleted-engine bug are not regressions and live with the other
+// RunPipeline behavior tests in run_test.go, not here.
 //
 // sdk v0.5.1 added ctx as Producer/Consumer's first parameter, giving
 // well-behaved plugins a way to exit on cancellation instead of parking.
@@ -267,49 +276,6 @@ func Test_GoroutinesDoNotAccumulateAcrossRuns(t *testing.T) {
 		baseline, runtime.NumGoroutine(), buf[:runtime.Stack(buf, true)])
 }
 
-// Capability test, not a regression: sdk v0.5.1 added ctx to Producer and
-// Consumer specifically so a plugin abandoned mid-send has a way to exit
-// instead of parking forever — the one leak PR #20's rewrite documented as
-// unavoidable ("the sdk contract has no context"). A producer that actually
-// selects on ctx.Done() alongside its send, cut off mid-stream by a
-// flow.Producer stop-after wrap, must leave no goroutine behind at all.
-func Test_CtxAwareProducer_LeavesNoGoroutineOnAbandon(t *testing.T) {
-	blockForever := func(ctx context.Context, send chan<- []byte, errs chan<- error) {
-		defer close(send)
-		defer close(errs)
-		for {
-			select {
-			case send <- []byte("x"):
-				continue
-			case <-ctx.Done():
-				return
-			}
-		}
-	}
-
-	baseline := runtime.NumGoroutine()
-	var got atomic.Int64
-	if err := panicSafeRun(t, &Pipeline{
-		Producers:   staticSource(flow.Producer(blockForever, 0, 0, 3)),
-		Parallel:    1,
-		Consumers:   []sdk.Consumer{countAll(&got)},
-		Transformer: sdk.Map(func(msg []byte) ([]byte, error) { return msg, nil }),
-	}); err != nil {
-		t.Fatal(err)
-	}
-
-	deadline := time.Now().Add(2 * time.Second)
-	for time.Now().Before(deadline) {
-		if runtime.NumGoroutine() <= baseline {
-			return
-		}
-		time.Sleep(20 * time.Millisecond)
-	}
-	buf := make([]byte, 1<<16)
-	t.Fatalf("ctx-aware producer still leaked a goroutine on abandon: %d -> %d\n%s",
-		baseline, runtime.NumGoroutine(), buf[:runtime.Stack(buf, true)])
-}
-
 // Regression for the late-error drop: a producer that closes its data
 // channel and then reports an error. The error forwarder used to sweep
 // pending errors with a non-blocking default the moment all data channels
@@ -344,74 +310,4 @@ func Test_ErrorAfterDataClose_IsDelivered(t *testing.T) {
 	if err == nil || !strings.Contains(err.Error(), "late error after data close") {
 		t.Fatalf("late error was dropped: got %v", err)
 	}
-}
-
-// A producer that ignores ctx entirely and sends forever with a bare send
-// (no select on ctx.Done) violates the sdk contract and leaks its own
-// goroutine when abandoned — the engine can't prevent that. What the engine
-// must still guarantee is that RunPipeline itself returns: the abandoned
-// plugin parks, the pipeline doesn't. The flow.Producer stop-after wrap is
-// what actually cuts the stream here — the misbehaving producer never would.
-func Test_ContractViolatingProducer_EngineStillReturns(t *testing.T) {
-	misbehaving := func(_ context.Context, send chan<- []byte, errs chan<- error) {
-		for {
-			send <- []byte("x") // bare send: parks forever once abandoned
-		}
-	}
-
-	var got atomic.Int64
-	if err := panicSafeRun(t, &Pipeline{
-		Producers:   staticSource(flow.Producer(misbehaving, 0, 0, 3)),
-		Parallel:    1,
-		Consumers:   []sdk.Consumer{countAll(&got)},
-		Transformer: sdk.Map(func(msg []byte) ([]byte, error) { return msg, nil }),
-	}); err != nil {
-		t.Fatal(err)
-	}
-	if n := got.Load(); n != 3 {
-		t.Fatalf("want 3 delivered, got %d", n)
-	}
-	// Note: this test intentionally leaks the misbehaving producer's
-	// goroutine — that's the documented cost of violating the contract.
-}
-
-// Same class as Test_ContractViolatingProducer_EngineStillReturns, but for
-// the transformer stage the channel-contract rewrite introduced: a
-// transformer that ignores ctx and blocks on a bare receive/send loop leaks
-// its own goroutine when abandoned, but RunPipeline itself must still
-// return — startTransform's fire-and-forget launch (see
-// core/transformstage.go) exists specifically so this can't hang the engine.
-func Test_ContractViolatingTransformer_EngineStillReturns(t *testing.T) {
-	misbehaving := func(_ context.Context, in <-chan []byte, out chan<- []byte, errs chan<- error) {
-		for msg := range in {
-			out <- msg // bare send: parks forever once abandoned
-		}
-	}
-
-	// The consumer finishing early is what abandons the transformer: the
-	// run loop stops draining stage.out, and the misbehaving transformer
-	// parks on its bare send.
-	var got atomic.Int64
-	consumer := func(_ context.Context, recv <-chan []byte, errs chan<- error, done chan<- struct{}) {
-		defer close(errs)
-		for range recv {
-			if got.Add(1) >= 3 {
-				close(done)
-				return
-			}
-		}
-	}
-	if err := panicSafeRun(t, &Pipeline{
-		Producers:   staticSource(emitForever([]byte("x"))),
-		Parallel:    1,
-		Consumers:   []sdk.Consumer{consumer},
-		Transformer: misbehaving,
-	}); err != nil {
-		t.Fatal(err)
-	}
-	if n := got.Load(); n != 3 {
-		t.Fatalf("want 3 delivered, got %d", n)
-	}
-	// Note: this test intentionally leaks the misbehaving transformer's
-	// goroutine — that's the documented cost of violating the contract.
 }
